aws_node/requirements.txt
langgraph==0.0.60
langchain-openai==0.1.8
python-dotenv==1.0.1
ipython==8.25.0
langchain==0.2.5
fastapi==0.111.0
boto3==1.34.128
uvicorn==0.25.0
pyyaml==6.0.1
pydantic==2.7.4
starlette==0.37.2